requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
statsmodels==0.14.0
scikit-learn==1.3.0
jupyter==1.0.0
ipykernel==6.25.0